\name{circle_area}
\alias{circle_area}
\title{Calculate the area of a circle}
\usage{
  circle_area(radius)
}
\arguments{
  \item{radius}{The radius of the circle.}
}
\description{
  This function calculates the area of a circle given its radius.
}
\examples{
  circle_area(5)
}

\name{circle_perimeter}
\alias{circle_perimeter}
\title{Calculate the perimeter of a circle}
\usage{
  circle_perimeter(radius)
}
\arguments{
  \item{radius}{The radius of the circle.}
}
\description{
  This function calculates the perimeter of a circle given its radius.
}
\examples{
  circle_perimeter(5)
}


