\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello()
}
\description{\name{circle_area}
\alias{circle_area}
\title{Calculate the area of a circle}
\usage{
circle_area(radius)
}
\arguments{
  \item{radius}{The radius of the circle.}
}
\description{
This function calculates the area of a circle given its radius.
}
\examples{
circle_area(5)
}

Prints 'Hello, world!'.
}
\examples{
hello()
}
