\name{rectangle_area}
\alias{rectangle_area}
\title{Calculate the area of a rectangle}
\usage{
  rectangle_area(length, width)
}
\arguments{
  \item{length}{The length of the rectangle.}
  \item{width}{The width of the rectangle.}
}
\description{
  This function calculates the area of a rectangle given its length and width.
}
\examples{
  rectangle_area(5, 3)
}


\name{rectangle_perimeter}
\alias{rectangle_perimeter}
\title{Calculate the perimeter of a rectangle}
\usage{
  rectangle_perimeter(length, width)
}
\arguments{
  \item{length}{The length of the rectangle.}
  \item{width}{The width of the rectangle.}
}
\description{
  This function calculates the perimeter of a rectangle given its length and width.
}
\examples{
  rectangle_perimeter(5, 3)
}
