\name{triangle_area}
\alias{triangle_area}
\title{Calculate the area of a triangle}
\usage{
  triangle_area(a, b, c)
}
\arguments{
  \item{a}{The length of the first side of the triangle.}
  \item{b}{The length of the second side of the triangle.}
  \item{c}{The length of the third side of the triangle.}
}
\description{
  This function calculates the area of a triangle given the lengths of its sides using Heron's formula.
}
\examples{
triangle_area(3, 4, 5)
}

\name{triangle_perimeter}
\alias{triangle_perimeter}
\title{Calculate the perimeter of a triangle}
\usage{
triangle_perimeter(a, b, c)
}
\arguments{
  \item{a}{The length of the first side of the triangle.}
  \item{b}{The length of the second side of the triangle.}
  \item{c}{The length of the third side of the triangle.}
}
\description{
This function calculates the perimeter of a triangle given the lengths of its sides.
}
\examples{
triangle_perimeter(3, 4, 5)
}
